requirements.txt
discord.py==2.3.2
requests==2.31.0
configobj==5.0.8
aiohappyeyeballs==2.6.1